requirements.txt
chromadb==0.4.22
sentence-transformers==2.2.2
pymupdf==1.23.21
Pillow==10.2.0
tqdm==4.66.1
torch==2.2.0
torchvision==0.17.0
numpy<2.0.0
transformers>=4.38.0
timm>=0.9.10
einops
ollama
